freebank: read the shared Core client
